+++ firmware/boron-ins-fsm/test/consoleFunctionTests.cpp
@@ -44,6 +44,14 @@ SCENARIO( "Turn_Debugging_Publishes_On_Off", "[toggle debug flag]" ) {
                 REQUIRE( returnVal == -1 );
             }
         }
+
+        WHEN("The function is called with 'e'") {
+            int returnVal = toggle_debugging_publishes("e");   
+
+            THEN("The function should return whether debugging publishes are turned on, in this case, false") {
+                REQUIRE(returnVal == 0);
+            }
+        }
     }
 
     GIVEN( "A true debug flag" ) {
@@ -82,6 +90,14 @@ SCENARIO( "Turn_Debugging_Publishes_On_Off", "[toggle debug flag]" ) {
 
             THEN( "the return value should be -1" ) {
                 REQUIRE( returnVal == -1 );
+            }
+        }
+
+        WHEN("The function is called with 'e'") {
+            int returnVal = toggle_debugging_publishes("e");   
+
+            THEN("The function should return whether debugging publishes are turned on, in this case, true") {
+                REQUIRE(returnVal == 1);
             }
         }
     }